requirements.txt
certifi==2018.4.16
chardet==3.0.4
Click==6.7
Flask==1.0.2
future==0.18.0
gunicorn==19.9.0
idna==2.6
itsdangerous==0.24
line-bot-sdk==1.14.0
MarkupSafe==1.0
requests==2.18.4
urllib3==1.22
Werkzeug==0.14.1